refactor: move model family
