+++ internal/agent/prompt/model_family_test.go
@@ -1,4 +1,4 @@
-package agent
+package prompt
 
 import (
 	"testing"
